requirements.txt
asgiref==3.9.2
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
Flask==2.0.1
fonttools==4.58.5
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.3.1
packaging==25.0
pillow==11.3.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
